'LPlot/LLines': Added flag 'bPeriod' to display x-axis in units of period.
Details: For 'bPeriod = TRUE' the x-axis is displayed in units of period
(inverse frequency) with increasing time scale to the left, so that the shape
of the spectra does not change compared to using frequencies. This commit also
includes insertion of 'xlim/ylim' parameters to 'LPlot' (defaulting to NULL)
which is necessary to ensure proper default scaling of the x-axis.